feat(footer):footer refactoring and bug fix
