+++ movie-app/src/scss/main.scss
@@ -8,7 +8,7 @@ html, body {
     rgba(239, 237, 229, 1) 25%
   );
   
-  min-height: 100%;
+  // min-height: 100vh;
   bottom:0;
 }
 * {
@@ -24,9 +24,19 @@ html, body {
 .App {
   display: flex;
   flex-direction: column;
+  justify-content: space-between;
+  min-height: 100vh;
+}
+
+.empty-state {
+  width: 100vw;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
 
   .error-text {
-    margin: 3rem 0 0.5rem 5%;
+    margin-bottom: 20px;
   }
 }
